Fix compile warnings etc

+++ code/minimal-gen-sets/rebuild.sh
@@ -2,5 +2,7 @@
 set -e
 
 cp test-bmat8-enum.cpp libsemigroups/tests
-cd libsemigroups && make test_bmat8_enum -j8
-ln -s libsemigroups/test_bmat8_enum test_bmat8_enum
+cd libsemigroups && make test_bmat8_enum -j8 && cd ..
+if [ ! -L test_bmat8_enum ]; then
+  ln -s libsemigroups/test_bmat8_enum test_bmat8_enum
+fi
